Project proposal: SCIoI Open TWIPR
I'm a researcher at a german research cluster (Science of Intelligence https://www.scienceofintelligence.de) and we are investigating collective learning in robotic systems. For this we are developing an open-source two-wheeled inverted pendulum robot, which should enable us and others to study complex, communicating dynamic systems. For this, we relied on Raspberry Pi CM4's, but your board seems more powerful and more suited for complex vision-based algorithms and the USB 3.0 will enable the robot to use more advanced cameras and sensors.