Add T6x-FSBoard (#41)
Drop-in replacement motherboard for 15.4 inch T61 / 14.1 inch 4:3 T61 + 14/15 inch T60.

Currently designed for Radxa CM3/CM5 only, because of the SATA and USB3 capability. This might change in the future.
First prototype (with limited capability) is close to getting sent out for PCB production.
The idea is to have an m.2 slot for WiFi, and use the built-in NIC for onboard ethernet.
Designing is around 40-50% done. Will open source the design (probably under the GPL) once the project is completed.
Once the Radxa CM5 comes out, I will test that on this project too.